requirements.txt
langchain-core
langchain-community
streamlit
python-dotenv
pypdf
faiss-cpu
groq
langchain-groq
langserve
wikipedia